move mapmodule plugin publisher tools to one place under mapmodule
