+++ bundles/mapping/mapmodule/publisher/indexmap/IndexMapTool.js
@@ -1,33 +1,25 @@
-import { AbstractPublisherTool } from '../../../../../framework/publisher2/tools/AbstractPublisherTool';
+import { AbstractPublisherTool } from '../../../../framework/publisher2/tools/AbstractPublisherTool';
 
-class ZoombarTool extends AbstractPublisherTool {
+class IndexMapTool extends AbstractPublisherTool {
     constructor (...args) {
         super(...args);
         this.index = 1;
         this.group = 'additional';
-    };
+    }
 
-    /**
-    * Get tool object.
-    * @method getTool
-    *
-    * @returns {Object} tool description
-    */
     getTool () {
         return {
-            id: 'Oskari.mapframework.bundle.mapmodule.plugin.Portti2Zoombar',
-            title: Oskari.getMsg('MapModule', 'publisherTools.Zoombar.toolLabel'),
+            id: 'Oskari.mapframework.bundle.mapmodule.plugin.IndexMapPlugin',
+            title: Oskari.getMsg('MapModule', 'publisherTools.IndexMapPlugin.toolLabel'),
             config: this.state.pluginConfig || {}
         };
-    };
+    }
 
-    /**
-    * Get values.
-    * @method getValues
-    * @public
-    *
-    * @returns {Object} tool value object
-    */
+    isDisplayed () {
+        // not shown on 3d maps
+        return !Oskari.getSandbox().getMap().getSupports3D();
+    }
+
     getValues () {
         if (!this.isEnabled()) {
             return null;
@@ -41,15 +33,15 @@ class ZoombarTool extends AbstractPublisherTool {
                 }
             }
         };
-    };
+    }
 }
 
 // Attach protocol to make this discoverable by Oskari publisher
-Oskari.clazz.defineES('Oskari.publisher.ZoombarTool',
-    ZoombarTool,
+Oskari.clazz.defineES('Oskari.publisher.IndexMapTool',
+    IndexMapTool,
     {
         protocol: ['Oskari.mapframework.publisher.Tool']
     }
 );
 
-export { ZoombarTool };
+export { IndexMapTool };
