+++ bundles/mapping/mapmodule/publisher/mylocation/MyLocationComponent.jsx
