+++ bundles/mapping/mapmodule/publisher/mylocation/MyLocationHandler.js
