+++ bundles/mapping/mapmodule/publisher/panbuttons/PanButtonsComponent.jsx
