+++ bundles/mapping/mapmodule/publisher/panbuttons/PanButtonsHandler.js
