+++ bundles/mapping/mapmodule/publisher/tools.js
@@ -1,11 +1,11 @@
 /**
  * Import plugins to be exposed to publisher
  */
-import { LogoTool } from '../plugin/logo/publisher/LogoTool';
-import { ScaleBarTool } from '../plugin/scalebar/publisher/ScalebarTool';
-import { MyLocationTool } from '../plugin/mylocation/publisher/MyLocationTool';
-import { PanButtonsTool } from '../plugin/panbuttons/publisher/PanButtonsTool';
-import { IndexMapTool } from '../plugin/indexmap/publisher/IndexMapTool';
-import { ZoombarTool } from '../plugin/zoombar/publisher/ZoombarTool';
+import { LogoTool } from './logo/LogoTool';
+import { ScaleBarTool } from './scalebar/ScalebarTool';
+import { MyLocationTool } from './mylocation/MyLocationTool';
+import { PanButtonsTool } from './panbuttons/PanButtonsTool';
+import { IndexMapTool } from './indexmap/IndexMapTool';
+import { ZoombarTool } from './zoombar/ZoombarTool';
 
 export { LogoTool, ScaleBarTool, MyLocationTool, PanButtonsTool, IndexMapTool, ZoombarTool };
